Save raw int value in token
It became difficult to know how many positions to jump forward in the string if
an INT was found, reason being is that the logic in the code passes the string
from the last position of a found token.

Since the value of the INT token is not the same length, the lexer was getting
confused and sending incorrect positions for tokeniser.

Instead save raw value and I will later create methods to convert to either an
integer or character.