Working Lexer
Finally have a working lexer that produces the correct tokens for each
IMP/command in whitespace. I have tested this against multiple ws scripts
(helloworld and count) and the correct tokens are being produced.

Quite a large number of changes, one of the main ones is a list of lists, where
each list item is an IMP containing the IMP, command and if required the
argument.

This, I am hoping will make the latter stages of the interpreter easier to
implement. As I have the basis for a stack.