Distinguish between labels and integers
The code previously was make no distinction between a label and integer, the
thinking being that I would treat the label as an integer. The problem being is
that integers are signed where as labels have no sign. As a result, some labels
were raising exceptions as less elements were found in the `found` list when
scanning the string.

Code updated to make a distinction between 'PUSH' and other types that have
ARGS.... 'PUSH' will return a signed integer token.